subprocess_check_output was defined 3 times in the codebase
