+++ shared/modules/ssgcommon.py
@@ -1,5 +1,6 @@
 import datetime
 import platform
+import subprocess
 
 
 xml_version = """<?xml version="1.0" encoding="UTF-8"?>"""
@@ -56,3 +57,29 @@ def oval_generated_header(product_name, schema_version, ssg_version):
         <oval:timestamp>%s</oval:timestamp>
     </generator>""" % (product_name, ssg_version, platform.python_version(),
                        schema_version, timestamp)
+
+
+def subprocess_check_output(*popenargs, **kwargs):
+    # Backport of subprocess.check_output taken from
+    # https://gist.github.com/edufelipe/1027906
+    #
+    # Originally from Python 2.7 stdlib under PSF, compatible with LGPL2+
+    # Copyright (c) 2003-2005 by Peter Astrand <astrand@lysator.liu.se>
+    # Changes by Eduardo Felipe
+
+    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
+    output, unused_err = process.communicate()
+    retcode = process.poll()
+    if retcode:
+        cmd = kwargs.get("args")
+        if cmd is None:
+            cmd = popenargs[0]
+        error = subprocess.CalledProcessError(retcode, cmd)
+        error.output = output
+        raise error
+    return output
+
+
+if hasattr(subprocess, "check_output"):
+    # if available we just use the real function
+    subprocess_check_output = subprocess.check_output
