+++ shared/utils/build-all-guides.py
@@ -15,8 +15,6 @@ except ImportError:
 
 import os.path
 import argparse
-import subprocess
-
 import threading
 import Queue
 import sys
@@ -33,32 +31,6 @@ import ssgcommon
 OSCAP_PATH = "oscap"
 
 
-def subprocess_check_output(*popenargs, **kwargs):
-    # Backport of subprocess.check_output taken from
-    # https://gist.github.com/edufelipe/1027906
-    #
-    # Originally from Python 2.7 stdlib under PSF, compatible with LGPL2+
-    # Copyright (c) 2003-2005 by Peter Astrand <astrand@lysator.liu.se>
-    # Changes by Eduardo Felipe
-
-    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
-    output, unused_err = process.communicate()
-    retcode = process.poll()
-    if retcode:
-        cmd = kwargs.get("args")
-        if cmd is None:
-            cmd = popenargs[0]
-        error = subprocess.CalledProcessError(retcode, cmd)
-        error.output = output
-        raise error
-    return output
-
-
-if hasattr(subprocess, "check_output"):
-    # if available we just use the real function
-    subprocess_check_output = subprocess.check_output
-
-
 def generate_guide_for_input_content(input_content, benchmark_id, profile_id):
     """Returns HTML guide for given input_content and profile_id
     combination. This function assumes only one Benchmark exists
@@ -72,9 +44,7 @@ def generate_guide_for_input_content(input_content, benchmark_id, profile_id):
         args.extend(["--profile", profile_id])
     args.append(input_content)
 
-    ret = subprocess_check_output(args).decode("utf-8")
-
-    return ret
+    return ssgcommon.subprocess_check_output(args).decode("utf-8")
 
 
 def get_cpu_count():
