+++ shared/utils/build-all-remediation-roles.py
@@ -14,7 +14,6 @@ except ImportError:
 
 import os.path
 import argparse
-import subprocess
 
 import threading
 import Queue
@@ -27,37 +26,13 @@ sys.path.insert(0, os.path.join(
         os.path.dirname(os.path.dirname(os.path.realpath(__file__))),
         "modules"))
 import xccdf_utils
+import ssgcommon
 
 OSCAP_PATH = "oscap"
 
 
-def subprocess_check_output(*popenargs, **kwargs):
-    # Backport of subprocess.check_output taken from
-    # https://gist.github.com/edufelipe/1027906
-    #
-    # Originally from Python 2.7 stdlib under PSF, compatible with LGPL2+
-    # Copyright (c) 2003-2005 by Peter Astrand <astrand@lysator.liu.se>
-    # Changes by Eduardo Felipe
-
-    process = subprocess.Popen(stdout=subprocess.PIPE, *popenargs, **kwargs)
-    output, unused_err = process.communicate()
-    retcode = process.poll()
-    if retcode:
-        cmd = kwargs.get("args")
-        if cmd is None:
-            cmd = popenargs[0]
-        error = subprocess.CalledProcessError(retcode, cmd)
-        error.output = output
-        raise error
-    return output
-
-
-if hasattr(subprocess, "check_output"):
-    # if available we just use the real function
-    subprocess_check_output = subprocess.check_output
-
-
-def generate_role_for_input_content(input_content, benchmark_id, profile_id, template):
+def generate_role_for_input_content(input_content, benchmark_id, profile_id,
+                                    template):
     """Returns remediation role for given input_content and profile_id
     combination. This function assumes only one Benchmark exists
     in given input_content!
@@ -74,9 +49,7 @@ def generate_role_for_input_content(input_content, benchmark_id, profile_id, tem
     args.extend(["--template", template])
     args.append(input_content)
 
-    ret = subprocess_check_output(args).decode("utf-8")
-
-    return ret
+    return ssgcommon.subprocess_check_output(args).decode("utf-8")
 
 
 def add_minimum_ansible_version(ansible_src):
